Add files via upload
#  Description and goal

In this assignment you will provide data analysis within the context of a business application. The 
specifications below indicate what you need to produce, but not how to produce it.
MoCo is an organisation investing in movies. Using a multitude of factors, the managers want to
• Have a better understanding and visualisation of the data through an interactive dashboard.
• In addition, the dashboard should also help understanding
o the factors influencing positively o or negatively the potential success or revenues of
future movie projects
o and the evolution of the market over the years

#  Data - The HOLLYWOODMOVIEDATASET_IM data set.

The data set, available in SAS Viya, was obtained from several movie databases using both automated 
as well as manual means. It is more than likely that some of the values are captured/entered 
incorrectly. Hence, the accuracy of the data set cannot be guaranteed. This data set can be used for 
descriptive and predictive modelling